asp_assignment_1/requirements.txt
mysql-connector-python==8.0.30
numpy==1.23.3
pandas==1.5.0
protobuf==3.20.1
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0